requirements.txt
streamlit>=1.35.0

pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23

sentence-transformers>=3.0.0

chromadb>=0.5.0

anthropic>=0.28.0
openai>=1.30.0

python-dotenv>=1.0.0
